+++ src/app/place/[id]/components/index.ts
@@ -1,0 +1,1 @@
+export { default as ReviewSection } from './ReviewSection';
