+++ src/app/place/components/Footer/index.tsx
@@ -1,4 +1,4 @@
-import TimeLogSheet from './TimeLogSheet';
+import TimeLogSheet from '../TimeLogSheet';
 import { Button } from '@/components';
 import { useState } from 'react';
 
@@ -10,7 +10,7 @@ export default function Footer() {
       <footer>
         <Button
           type="DEFAULT"
-          className="fixed bottom-0 z-50 w-full min-w-[360px] max-w-[448px]"
+          className="fixed inset-x-0 bottom-0 z-50 mx-auto w-full min-w-[360px] max-w-[448px]"
           onClick={() => setIsLogTimeSheetOpen(true)}
         >
           카공 기록하기
