+++ src/app/place/components/ReviewSheet/index.tsx
@@ -78,7 +78,7 @@ export default function ReviewSheet({ isOpen, onClose }: ReviewSheetProps) {
       {isOpen && (
         <Button
           type="DEFAULT"
-          className="fixed bottom-0 z-[1000] w-full min-w-[360px] max-w-[448px]"
+          className="fixed inset-x-0 bottom-0 z-[60] mx-auto w-full min-w-[360px] max-w-[448px]"
           disabled
         >
           리뷰등록
