+++ src/app/place/components/index.ts
@@ -2,9 +2,9 @@ export { default as Header } from './Header';
 export { default as KagongItem } from './KagongItem';
 export { default as ReviewItem } from './ReviewItem';
 export { default as Tag } from './Tag';
-export { default as Tooltip } from './Tooltip';
+export { default as Tooltip } from '../[id]/components/ReviewSection';
 export { default as Info } from './Info';
 export { default as ReviewSheet } from './ReviewSheet';
-export { default as TimeLogSheet } from './Footer/TimeLogSheet';
+export { default as TimeLogSheet } from './TimeLogSheet';
 export { default as ShareSheet } from './ShareSheet';
 export { default as PlaceTopNavigationBar } from './PlaceTopNavigationBar';
